+++ Network/layers/MaxPoolingLayer.h
@@ -14,8 +14,8 @@ namespace cn {
         std::optional<Bitmap<Vector2<int>>> mapping;
 
         MaxPoolingLayer(int _id, Network &_network, int _kernelSizeX, int _kernelSizeY);
-        Bitmap<float> run(const Bitmap<float> &input) override;
-        float getChain(const Vector3<int> &inputPos) override;
+        Bitmap<double> run(const Bitmap<double> &input) override;
+        double getChain(const Vector3<int> &inputPos) override;
     };
 }
