+++ Network/layers/OutputLayer.cpp
@@ -7,15 +7,15 @@
 
 cn::OutputLayer::OutputLayer(int id, cn::Network &network) : FlatteningLayer(id, network) {}
 
-cn::Bitmap<float> cn::OutputLayer::run(const cn::Bitmap<float> &input) {
+cn::Bitmap<double> cn::OutputLayer::run(const cn::Bitmap<double> &input) {
     return FlatteningLayer::run(input);
 }
 
-float cn::OutputLayer::getChain(const Vector3<int> &input) {
+double cn::OutputLayer::getChain(const Vector3<int> &input) {
     return network->getOutput(__id)->getCell(input) - target->getCell(input);
 }
 
-void cn::OutputLayer::setTarget(const cn::Bitmap<float> *_target) {
+void cn::OutputLayer::setTarget(const cn::Bitmap<double> *_target) {
     target = _target;
     if(outputSize != target->size()){
         throw std::logic_error("");
