+++ Network/layers/interfaces/Learnable.h
@@ -15,15 +15,15 @@ namespace cn {
         int getNeuronsCount() const;
         virtual int weightsCount() const = 0;
         virtual int biasesCount() const = 0;
-        virtual void setWeight(int weightID, float value) = 0;
-        virtual float getWeight(int weightID) const = 0;
-        virtual void setBias(int neuronID, float value) = 0;
-        virtual float getBias(int neuronID) const = 0;
+        virtual void setWeight(int weightID, double value) = 0;
+        virtual double getWeight(int weightID) const = 0;
+        virtual void setBias(int neuronID, double value) = 0;
+        virtual double getBias(int neuronID) const = 0;
 
         virtual void randomInit() = 0;
         Learnable(int id, Network &network, int neuronsCount);
-        virtual std::vector<float> getWeightsGradient() = 0;
-        virtual std::vector<float> getBiasesGradient() = 0;
+        virtual std::vector<double> getWeightsGradient() = 0;
+        virtual std::vector<double> getBiasesGradient() = 0;
     };
 }
