+++ Utils/Differentiables/Sigmoid.cpp
@@ -5,11 +5,11 @@
 #include "Sigmoid.h"
 #include <cmath>
 
-float Sigmoid::func(float x) const {
+double Sigmoid::func(double x) const {
     return 1.f/(1.f + std::pow(std::exp(1), -x));
 }
 
-float Sigmoid::derive(float x) const {
-    float sig = func(x);
+double Sigmoid::derive(double x) const {
+    double sig = func(x);
     return sig * (1.f - sig);
 }
