+++ Utils/dataStructures/PointData.cpp
@@ -3,9 +3,9 @@
 //
 #include "PointData.h"
 
-PointData::PointData(const std::pair<float, float> &_point, void * _data): point(_point), data(_data)
+PointData::PointData(const std::pair<double, double> &_point, void * _data): point(_point), data(_data)
 {}
 
-PointData::PointData(const std::pair<float, float> &_point): point(_point){
+PointData::PointData(const std::pair<double, double> &_point): point(_point){
 
 }
